requirements.txt
Flask>=2.2.3
mysql-connector-python==9.0.0
gunicorn==23.0.0
python-dotenv==1.0.1
